# backend/app/generated/main.py
# generated by fastapi-codegen:
#   filename:  openapi.json
#   timestamp: 2025-01-07T02:31:32+00:00

from __future__ import annotations

from typing import Any, Union

from fastapi import FastAPI, UploadFile

from .models import (
    HTTPValidationError,
    OpenAIKeyUpdate,
    ProfileUpdate,
    ScanRequest,
    TokenRequest,
    TokenResponse,
)

app = FastAPI(
    title='Security Scanner API',
    description='API for security scanning and compliance monitoring',
    version='1.0.0',
)


@app.get('/api/ai/models', response_model=Any)
def get_ai_models_api_ai_models_get() -> Any:
    """
    Get Ai Models
    """
    pass


@app.get('/api/ai/recommendations', response_model=Any)
def get_ai_recommendations_api_ai_recommendations_get() -> Any:
    """
    Get Ai Recommendations
    """
    pass


@app.post(
    '/api/analyze',
    response_model=Any,
    responses={'422': {'model': HTTPValidationError}},
)
def analyze_scan_api_analyze_post(body: ScanRequest) -> Union[Any, HTTPValidationError]:
    """
    Analyze Scan
    """
    pass


@app.get('/api/compliance/trends', response_model=Any)
def get_compliance_trends_api_compliance_trends_get() -> Any:
    """
    Get Compliance Trends
    """
    pass


@app.get('/api/documents', response_model=Any)
def get_documents_api_documents_get() -> Any:
    """
    Get Documents
    """
    pass


@app.post(
    '/api/documents',
    response_model=Any,
    responses={'422': {'model': HTTPValidationError}},
)
def upload_document_api_documents_post(
    file: UploadFile,
) -> Union[Any, HTTPValidationError]:
    """
    Upload Document
    """
    pass


@app.get(
    '/api/documents/{doc_id}',
    response_model=Any,
    responses={'422': {'model': HTTPValidationError}},
)
def get_document_api_documents__doc_id__get(
    doc_id: int,
) -> Union[Any, HTTPValidationError]:
    """
    Get Document
    """
    pass


@app.delete(
    '/api/documents/{doc_id}',
    response_model=Any,
    responses={'422': {'model': HTTPValidationError}},
)
def delete_document_api_documents__doc_id__delete(
    doc_id: int,
) -> Union[Any, HTTPValidationError]:
    """
    Delete Document
    """
    pass


@app.get('/api/embeddings', response_model=Any)
def get_embeddings_api_embeddings_get() -> Any:
    """
    Get Embeddings
    """
    pass


@app.get(
    '/api/placeholder/{width}/{height}',
    response_model=Any,
    responses={'422': {'model': HTTPValidationError}},
)
def get_placeholder_image_api_placeholder__width___height__get(
    width: int, height: int = ...
) -> Union[Any, HTTPValidationError]:
    """
    Get Placeholder Image
    """
    pass


@app.get('/api/protected', response_model=Any)
def protected_route_api_protected_get() -> Any:
    """
    Protected Route
    """
    pass


@app.get('/api/scans', response_model=Any)
def get_scans_api_scans_get() -> Any:
    """
    Get Scans
    """
    pass


@app.get(
    '/api/scans/{scan_id}',
    response_model=Any,
    responses={'422': {'model': HTTPValidationError}},
)
def get_scan_api_scans__scan_id__get(scan_id: int) -> Union[Any, HTTPValidationError]:
    """
    Get Scan
    """
    pass


@app.delete(
    '/api/scans/{scan_id}',
    response_model=Any,
    responses={'422': {'model': HTTPValidationError}},
)
def delete_scan_api_scans__scan_id__delete(
    scan_id: int,
) -> Union[Any, HTTPValidationError]:
    """
    Delete Scan
    """
    pass


@app.get('/api/statistics', response_model=Any)
def get_statistics_api_statistics_get() -> Any:
    """
    Get Statistics
    """
    pass


@app.get('/api/tokens', response_model=Any)
def list_service_tokens_api_tokens_get() -> Any:
    """
    List Service Tokens
    """
    pass


@app.post(
    '/api/tokens',
    response_model=TokenResponse,
    responses={'422': {'model': HTTPValidationError}},
)
def create_service_token_api_tokens_post(
    body: TokenRequest,
) -> Union[TokenResponse, HTTPValidationError]:
    """
    Create Service Token
    """
    pass


@app.delete(
    '/api/tokens/{token_id}',
    response_model=Any,
    responses={'422': {'model': HTTPValidationError}},
)
def revoke_token_api_tokens__token_id__delete(
    token_id: int,
) -> Union[Any, HTTPValidationError]:
    """
    Revoke Token
    """
    pass


@app.get('/api/user/has-openai-key', response_model=Any)
def check_openai_key_api_user_has_openai_key_get() -> Any:
    """
    Check Openai Key
    """
    pass


@app.post(
    '/api/user/openai-key',
    response_model=Any,
    responses={'422': {'model': HTTPValidationError}},
)
def update_openai_key_api_user_openai_key_post(
    body: OpenAIKeyUpdate,
) -> Union[Any, HTTPValidationError]:
    """
    Update Openai Key
    """
    pass


@app.get('/api/user/profile', response_model=Any)
def get_user_profile_api_user_profile_get() -> Any:
    """
    Get User Profile
    """
    pass


@app.patch(
    '/api/user/profile',
    response_model=Any,
    responses={'422': {'model': HTTPValidationError}},
)
def update_user_profile_api_user_profile_patch(
    body: ProfileUpdate,
) -> Union[Any, HTTPValidationError]:
    """
    Update User Profile
    """
    pass


@app.get('/api/vulnerability-distribution', response_model=Any)
def get_vulnerability_distribution_api_vulnerability_distribution_get() -> Any:
    """
    Get Vulnerability Distribution
    """
    pass