# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Security Scanner API",
        "description": "API for security scanning and compliance monitoring",
        "version": "1.0.0"
    },
    "paths": {
        "/api/tokens": {
            "get": {
                "summary": "List Service Tokens",
                "description": "List all service account tokens",
                "operationId": "list_service_tokens_api_tokens_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create Service Token",
                "description": "Create a new service account token",
                "operationId": "create_service_token_api_tokens_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TokenRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tokens/{token_id}": {
            "delete": {
                "summary": "Revoke Token",
                "description": "Revoke a service account token",
                "operationId": "revoke_token_api_tokens__token_id__delete",
                "parameters": [
                    {
                        "name": "token_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Token Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/analyze": {
            "post": {
                "summary": "Analyze Scan",
                "operationId": "analyze_scan_api_analyze_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ScanRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/scans/{scan_id}": {
            "get": {
                "summary": "Get Scan",
                "operationId": "get_scan_api_scans__scan_id__get",
                "parameters": [
                    {
                        "name": "scan_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Scan Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Scan",
                "operationId": "delete_scan_api_scans__scan_id__delete",
                "parameters": [
                    {
                        "name": "scan_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Scan Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/documents": {
            "get": {
                "summary": "Get Documents",
                "operationId": "get_documents_api_documents_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Upload Document",
                "operationId": "upload_document_api_documents_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_document_api_documents_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/documents/{doc_id}": {
            "get": {
                "summary": "Get Document",
                "operationId": "get_document_api_documents__doc_id__get",
                "parameters": [
                    {
                        "name": "doc_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Doc Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Document",
                "operationId": "delete_document_api_documents__doc_id__delete",
                "parameters": [
                    {
                        "name": "doc_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Doc Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/placeholder/{width}/{height}": {
            "get": {
                "summary": "Get Placeholder Image",
                "description": "Generate a placeholder image with the specified dimensions",
                "operationId": "get_placeholder_image_api_placeholder__width___height__get",
                "parameters": [
                    {
                        "name": "width",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Width"
                        }
                    },
                    {
                        "name": "height",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Height"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/profile": {
            "get": {
                "summary": "Get User Profile",
                "description": "Get user profile information",
                "operationId": "get_user_profile_api_user_profile_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            },
            "patch": {
                "summary": "Update User Profile",
                "description": "Update user profile information",
                "operationId": "update_user_profile_api_user_profile_patch",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/user/openai-key": {
            "post": {
                "summary": "Update Openai Key",
                "description": "Update user's OpenAI API key",
                "operationId": "update_openai_key_api_user_openai_key_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OpenAIKeyUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/user/has-openai-key": {
            "get": {
                "summary": "Check Openai Key",
                "description": "Check if user has stored an OpenAI API key",
                "operationId": "check_openai_key_api_user_has_openai_key_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/protected": {
            "get": {
                "summary": "Protected Route",
                "operationId": "protected_route_api_protected_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/ai/recommendations": {
            "get": {
                "summary": "Get Ai Recommendations",
                "description": "Get AI-generated recommendations based on recent scans and compliance data",
                "operationId": "get_ai_recommendations_api_ai_recommendations_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/ai/models": {
            "get": {
                "summary": "Get Ai Models",
                "description": "Get available AI models and their status",
                "operationId": "get_ai_models_api_ai_models_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/compliance/trends": {
            "get": {
                "summary": "Get Compliance Trends",
                "description": "Get compliance trends over time",
                "operationId": "get_compliance_trends_api_compliance_trends_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/statistics": {
            "get": {
                "summary": "Get Statistics",
                "description": "Get overall statistics for the dashboard",
                "operationId": "get_statistics_api_statistics_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Body_upload_document_api_documents_post": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "category": {
                        "type": "string",
                        "title": "Category"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "version": {
                        "type": "string",
                        "title": "Version",
                        "default": "1.0"
                    },
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "category",
                    "description",
                    "file"
                ],
                "title": "Body_upload_document_api_documents_post"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "OpenAIKeyUpdate": {
                "properties": {
                    "openai_key": {
                        "type": "string",
                        "title": "Openai Key"
                    }
                },
                "type": "object",
                "required": [
                    "openai_key"
                ],
                "title": "OpenAIKeyUpdate"
            },
            "ProfileUpdate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "jobTitle": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Jobtitle"
                    },
                    "department": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Department"
                    }
                },
                "type": "object",
                "required": [
                    "name"
                ],
                "title": "ProfileUpdate"
            },
            "ScanRequest": {
                "properties": {
                    "project_name": {
                        "type": "string",
                        "title": "Project Name"
                    },
                    "trivy_data": {
                        "type": "object",
                        "title": "Trivy Data"
                    },
                    "openai_key": {
                        "type": "string",
                        "title": "Openai Key"
                    }
                },
                "type": "object",
                "required": [
                    "project_name",
                    "trivy_data",
                    "openai_key"
                ],
                "title": "ScanRequest"
            },
            "TokenRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "expires_in_days": {
                        "type": "integer",
                        "title": "Expires In Days",
                        "default": 30
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "description"
                ],
                "title": "TokenRequest"
            },
            "TokenResponse": {
                "properties": {
                    "token": {
                        "type": "string",
                        "title": "Token"
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Expires At"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    }
                },
                "type": "object",
                "required": [
                    "token",
                    "expires_at",
                    "id",
                    "name"
                ],
                "title": "TokenResponse"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}